+++ packages/terra-theme-provider/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 4.5.0 - (August 27, 2020)
+
+* Fixed
+  * Update wdio screenshots for fusion theme that depended on terra-dev-site theme values that are now scoped.
+
 ## 4.4.0 - (August 4, 2020)
 
 * Changed